handle upload

+++ app/src/main/java/com/gianghv/android/domain/AppState.kt
@@ -2,7 +2,7 @@ package com.gianghv.android.domain
 
 object AppState {
     var logined: Boolean = false
-    var userId: Int = 1
+    var userId: Long = 1
     var name: String = ""
     var firstAppOpen: Boolean = false
     var userRole: UserRole = UserRole.RESEARCHER
